+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Mapbox welcomes participation and contributions from everyone.
 
+* Use MapboxCoreMaps API to move a Layer instead of manually removing the layer then adding it back. ([#1367](https://github.com/mapbox/mapbox-maps-ios/pull/1367))
+
 ## 10.6.0-rc.1 - June 2, 2022
 
 * Update to MapboxCoreMaps 10.6.0-rc.1 and MapboxCommon 22.0.0-rc.2. ([#1368](https://github.com/mapbox/mapbox-maps-ios/pull/1368))